// src/app/payments/paypal/model/paypal.model.ts
export interface IPayPalConfig {

    /**
     * Currency - Defaults to USD if not provided
     */
    currency?: string;

    /**
    * Use when creating order on client
    */
    createOrderOnClient?: (data: any) => ICreateOrderRequest;

    /**
     * Use for creating orders on server. PayPal expects you to return 'orderId' in this method
     */
    createOrderOnServer?: (data: any) => Promise<string>;

    /**
     * Advanced configuration
     */
    advanced?: IAdvancedConfiguration;

    /**
     * Client id
     */
    clientId: string;

    /**
     * Shipping callback
     * see https://developer.paypal.com/docs/checkout/integration-features/shipping-callback/
     */
    onShippingChange?: OnShippingChangeCallback;

    /**
     * Called when 'onApprove' event occurs
     */
    onApprove: (data: IOnApproveCallbackData, actions: any) => void;

    /**
    * Called when authorization on client succeeds
    */
    onClientAuthorization?: (authorization: IClientAuthorizeCallbackData) => void;

    /**
     * Implement for authorizing on server side
     */
    authorizeOnServer?: (data: IOnApproveCallbackData, actions: any) => Promise<any>;

    /**
     * Button style configuration
     */
    style?: IPayPalButtonStyle;

    /**
     * Error handler
     */
    onError?: (err: any) => void;

    /**
     * Click handler
     */
    onClick?: () => void;

    /**
     * Cancel handler
     */
    onCancel?: (data: ICancelCallbackData, actions: any) => void;
}

export interface IClientAuthorizeCallbackData extends IOrderDetails {
    links: ILinkDescription[];
}

export interface IOrderDetails {
    create_time: string;
    update_time: string;
    id: string;
    intent: OrderIntent;
    payer: IPayer;
    status: OrderStatus;
    links: ILinkDescription[];
}

export interface ILinkDescription {
    href: string;
    rel: String;
    method?: LinkMethod;
}

export interface IQueryParam {
    name: string;
    value: string;
}

export type OnShippingChangeCallback = (data: IOnShippingChangeData, actions: IOnShippingChangeActions) => any;

export interface IOnShippingChangeData {
    paymentToken: string;
    shipping_address: any;
    selected_shipping_method?: any;
}

export interface IOnShippingChangeActions {
    resolve: () => any;
    reject: () => any;
    patch: () => any;
}

export type LinkMethod = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'HEAD' | 'CONNECT' | 'OPTIONS' | 'PATCH';

export interface IAdvancedConfiguration {
    updateOrderDetails?: {
        commit?: boolean
    };
    extraQueryParams?: IQueryParam[];
}

export interface IOnApproveCallbackData {
    orderID: string;
    payerID: string;
}

export interface ICreateOrderCallbackActions {
    order: {
        create: (order: ICreateOrderRequest) => Promise<any>;
    };
}

export interface ICancelCallbackData {
    orderID: string;
}

export interface IOnApproveCallbackActions {
    redirect: () => void;
    restart: () => void;
    order: {
        authorize: () => Promise<any>;
        capture: () => Promise<any>;
        get: () => Promise<IOrderDetails>;
        patch: () => Promise<any>;
    };
}

export interface IPayPalButtonStyle {
    label?: 'paypal' | 'checkout' | 'pay' | 'installment';
    size?: 'small' | 'medium' | 'large' | 'responsive';
    shape?: 'pill' | 'rect';
    color?: 'gold' | 'blue' | 'silver';
    layout?: 'horizontal' | 'vertical';
    tagline?: boolean;
}

export interface ICreateOrderRequest {
    intent: OrderIntent;
    purchase_units: IPurchaseUnit[];

    payer?: IPayer;
    application_context?: IApplicationContext;

}
export interface IPayer {
    name?: IPartyName;
    email_address?: string;
    payer_id?: string;
    birth_date?: string;
    tax_info?: ITaxInfo;
    address?: IAddressPortable;
}

export interface IApplicationContext {
    brand_name?: string;
    locale?: string;
    landing_page?: PaypalLandingPage;
    shipping_preference?: ShippingPreference;
    user_action?: PayPalUserAction;
    payment_method?: IPaymentMethod;
    return_url?: string;
    cancel_url?: string;
}
// ...
export type OrderIntent = 'CAPTURE' | 'AUTHORIZE';
// ...
export interface IPurchaseUnit {
    amount: IUnitAmount;

    reference_id?: string;
    payee?: IPayee;
    payment_instruction?: IPaymentInstructions;
    description?: string;
    custom_id?: string;
    invoice_id?: string;
    soft_descriptor?: string;
    items: ITransactionItem[];
    shipping?: IShipping;
}